+++ src/typescript/index.iife.tsx
@@ -4,8 +4,8 @@ import { StormPlayer } from "./StormPlayer";
 import { StormLibraryConfig } from "@stormstreaming/stormlibrary";
 import StormPlayerCustomElement from "./StormPlayerElement";
 
-export default function(playerConfig:StormPlayerConfig, libraryConfig:StormLibraryConfig):StormPlayer{
-    return new StormPlayer(playerConfig, libraryConfig);
+export default function(playerConfig:StormPlayerConfig, libraryConfig:StormLibraryConfig, wait:boolean):StormPlayer{
+    return new StormPlayer(playerConfig, libraryConfig, wait);
 }
 
 customElements.define("storm-player", StormPlayerCustomElement);